Refactoring LikelihoodSearchHandler
