+++ traffic_stops/static/js/app/states/md/LikelihoodOfSearch.js
@@ -3,6 +3,8 @@ import VisualBase from '../../base/VisualBase.js';
 import TableBase from '../../base/TableBase.js';
 import Stops from './defaults.js';
 
+import * as C from '../../common/LikelihoodOfSearch.js';
+
 import _ from 'underscore';
 import d3 from 'd3';
 import Backbone from 'backbone';
@@ -10,56 +12,9 @@ import $ from 'jquery';
 
 Backbone.$ = $;
 
-var LikelihoodSearchHandler = DataHandlerBase.extend({
-  clean_data: function () {
-
-    var years,
-        raw = this.get("raw_data");
-
-    // get available years
-    years = d3.set(raw.stops.map((v) => v.year)).values();
-    years.filter((v) => (v >= Stops.start_year));
-    years.push("Total");
-
-    // get total searches/stops for all years by purpose
-    var getTotals = (arr) => {
-      // calculate total for all years by purpose; push to array
-      var purposes = d3.nest()
-                       .key((d) => d.purpose)
-                       .entries(arr);
-
-      purposes.forEach((v) => {
-        // create new totals object, and reset ethnicity-values
-        var total = _.clone(v.values[0]);
-
-        _.keys(total).forEach((key) => {
-          if (Stops.ethnicities.indexOf(key) >= 0) {
-            total[key] = 0;
-          }
-        });
-
-        // sum data from all years
-        v.values.forEach((year) => {
-          _.keys(year).forEach((key) => {
-            if (Stops.ethnicities.indexOf(key) >= 0) {
-              total[key] += year[key];
-            }
-          });
-        });
-        total["year"] = "Total";
-        arr.push(total);
-      });
-    };
-
-    if (raw.stops.length > 0) getTotals(raw.stops);
-    if (raw.searches.length > 0) getTotals(raw.searches);
-
-    // set cleaned-data to handler
-    this.set("data", {
-      years: years,
-      raw: raw
-    });
-  }
+var LikelihoodSearchHandler = C.LikelihoodSearchHandlerBase.extend({
+  types: [Stops.ethnicities],
+  defaults: Stops
 });
 
 var LikelihoodOfSearch = VisualBase.extend({
